+++ CHANGELOG.rst
@@ -3,6 +3,11 @@
 ===========
 
 
+NEW CHANGES
+
+* Fix broken import in homely.install
+
+
 Version 0.11.0 - 16 Oct 2016
 ----------------------------